NLP Extractor now indexes on-the-fly
While each thread is being indexed, the documents go through the solr indexing code.  While the solr document is being created, NLP fields are being added.

The NLP fields take up a ton of processing time, we are caching the results in MapDB.

There is going to be a major refactor to use kafka instead.