changing to protocol buggers part 2:
Still in progress.  Discovered that the Protobuf version of serde's are a bit off.  When sending the messages back it seems we have to specify the type and somehow when the tests start, we should find a way to register the protobuf.  During testing it's returning only dynamic messages but suspect in a running container with confluent platform we will see the specific types return.  it's a pain

Another note: when working with the micronaut docker container it's a bit fussy with the following parameters:
1) Kafka client - can't be latest (avro allowed this)
2) if it's failing for "no reason" and it all looks good, delete the ~/.micronaut directory which is where the testing platform runs.  It seems like the files are easily corrupted.