First run of the semantic indexer
Right now this takes source documents from any solr instance, regardless of version as long as it can output a JSON response.

Then it will parse the json response as solr documents and insert them into the new solr 9 instance.

The next step is to create the document chunks and save the results to another collection that is chunked up.